Move relative-date-control to utils
